Test for Edit completed
